requirements.txt
ollama
requests
python-dotenv
flake8
autopep8
mypy
